Move Form stories and Checkbox stories into their own folder
